Strong types

+++ src/EditorFeatures/CSharpTest/Completion/CompletionProviders/FunctionPointerUnmanagedCallingConventionCompletionProviderTests.cs
@@ -72,6 +72,6 @@ public class FunctionPointerUnmanagedCallingConventionCompletionProviderTests : 
                 </Project>
             </Workspace>
             """;
-        await VerifyItemExistsAsync(markup, callingConvention, glyph: (int)Glyph.Keyword);
+        await VerifyItemExistsAsync(markup, callingConvention, glyph: Glyph.Keyword);
     }
 }
